7th task finally

+++ lab01/bit_ops.cpp
@@ -1,10 +1,10 @@
 #include "bit_ops.h"
-#include "stdint.h"
+
 
 // Возвращает бит номер n числа x.
 // Предполагается 0 <= n <= 31
-unsigned short get_bit(unsigned short x,
-                 unsigned short n) {
+unsigned get_bit(unsigned x,
+                 unsigned n) {
     return (x & 1 << n) >> n;
 }
 // Выставляет значение v в бите номер n
@@ -12,16 +12,16 @@ unsigned short get_bit(unsigned short x,
 /*  *x |= (v << n)  не проходит по двум тестам, 
 если подаётся бит нуля, то мы не сможем его установить, если бит n единица, 
 поэтому изначально мы обнуляем бит n*/ 
-void set_bit(unsigned short * x,
-             unsigned short n,
-             unsigned short v) {  
+void set_bit(unsigned * x,
+             unsigned n,
+             unsigned v) {  
     *x = (*x & ~ (1<<n)) | (v << n); 
 }
 
 // Меняет значение бита номер n на противоположное.
 // Предполагается что 0 <= n <= 31
-void flip_bit(unsigned short * x,
-              unsigned short n) {
+void flip_bit(unsigned * x,
+              unsigned n) {
     *x ^= 1 << n;
 }
